+++ relm4-macros/src/factory/mod.rs
@@ -1,5 +1,5 @@
 use proc_macro2::{Span as Span2, TokenStream as TokenStream2};
-use quote::{quote, quote_spanned, ToTokens};
+use quote::{quote, ToTokens};
 use syn::visit_mut::VisitMut;
 use syn::{parse_quote, Ident};
 
@@ -201,14 +201,9 @@ pub(crate) fn generate_tokens(
 
     let errors = errors.iter().map(syn::Error::to_compile_error);
 
-    let async_trait = asyncness.map(
-        |async_token| quote_spanned!(async_token.span => #[relm4::async_trait::async_trait(?Send)]),
-    );
-
     quote! {
         #widgets_struct
 
-        #async_trait
         #factory_impl
 
         #(#errors)*